# openapi/api.yaml
openapi: 3.0.3
info:
  title: RevU API
  description: RevU API for businesses and customers.
  version: 1.0.2
servers:
  - url: 'http://34.209.63.8:5000/api/v1'
    description: Production
  - url: 'http://localhost:5000/api/v1'
    description: Development
security:
  - bearerAuth: []
paths:
...
  /files:
    post:
      summary: Create a file
      tags:
        - files
      operationId: createFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  $ref: '#/components/schemas/MediumText'
                type:
                  allOf:
                    - $ref: '#/components/schemas/FileType'
                  default: image
                file:
                  type: string
                  format: binary
              required:
                - title
                - file
            examples:
              file:
                $ref: '#/components/examples/file'
      responses:
        '201':
          description: File created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/File'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    get:
      summary: Get files list
      tags:
        - files
      operationId: getFiles
      parameters:
        - name: type
          description: Limit result set to specific type
          in: query
          schema:
            $ref: '#/components/schemas/FileType'
        - $ref: '#/components/parameters/mimeTypes'
        - $ref: '#/components/parameters/paginationInclude'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/perPage'
      responses:
        '200':
          description: Paginated files list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedListSuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/File'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    delete:
      summary: Delete files
      tags:
        - files
      operationId: deleteFiles
      parameters:
        - $ref: '#/components/parameters/deleteIds'
      responses:
        '200':
          description: Files deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiObjectOperationSuccessResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /files/{id}:
    get:
      summary: Get a file
      tags:
        - files
      operationId: getFile
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: File details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/File'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
  /merchants:
    post:
      summary: Create a merchant
      tags:
        - merchants
      operationId: createMerchant
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Merchant'
            examples:
              merchant:
                $ref: '#/components/examples/merchant'
      responses:
        '201':
          description: Merchant created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Merchant'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    get:
      summary: Get merchants list
      tags:
        - merchants
      operationId: getMerchants
      parameters:
        - $ref: '#/components/parameters/merchantIncludeWithPagination'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/perPage'
      responses:
        '200':
          description: Paginated merchants list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedListSuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Merchant'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    delete:
      summary: Delete merchants
      tags:
        - merchants
      operationId: deleteMerchants
      parameters:
        - $ref: '#/components/parameters/deleteIds'
      responses:
        '200':
          description: Merchants deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiObjectOperationSuccessResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /merchants/{id}:
    get:
      summary: Get a merchant
      tags:
        - merchants
      operationId: getMerchant
      parameters:
        - $ref: '#/components/parameters/idIncludingZero'
        - $ref: '#/components/parameters/merchantInclude'
      responses:
        '200':
          description: Merchant details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Merchant'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
  /orders/{id}/items:
    post:
      summary: Create an orderItem
      tags:
        - orderItems
      operationId: createOrderItem
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItem'
            examples:
              orderItem:
                $ref: '#/components/examples/orderItem'
      responses:
        '201':
          description: OrderItem created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OrderItem'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    get:
      summary: Get orderItems list
      tags:
        - orderItems
      operationId: getOrderItems
      parameters:
        - $ref: '#/components/parameters/id'
        - $ref: '#/components/parameters/optedServicesIds'
        - $ref: '#/components/parameters/employeeId'
      responses:
        '200':
          description: Paginated orderItems list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedListSuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/OrderItem'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    put:
      summary: Update orderItems
      tags:
        - orderItems
      operationId: updateOrderItems
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/OrderItem'
      responses:
        '200':
          description: OrderItems created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ListSuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/OrderItem'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    delete:
      summary: Delete orderItems
      tags:
        - orderItems
      operationId: deleteOrderItems
      parameters:
        - $ref: '#/components/parameters/id'
        - $ref: '#/components/parameters/deleteIds'
      responses:
        '200':
          description: OrderItems deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiObjectOperationSuccessResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /orders/{orderId}/items/{id}:
    get:
      summary: Get an orderItem
      tags:
        - orderItems
      operationId: getOrderItem
      parameters:
        - $ref: '#/components/parameters/id'
        - $ref: '#/components/parameters/pathOrderId'
      responses:
        '200':
          description: OrderItem details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OrderItem'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    patch:
      summary: Update an orderItem
      tags:
        - orderItems
      operationId: updateOrderItem
      parameters:
        - $ref: '#/components/parameters/id'
        - $ref: '#/components/parameters/pathOrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItemUpdateBody'
      responses:
        '200':
          description: OrderItem updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OrderItem'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
components:
  parameters:
    id:
      name: id
      description: Resource id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/IntegerId'
    idIncludingZero:
      name: id
      description: Resource id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/PositiveIntegerIncludingZero'
...
    deleteIds:
      name: ids
      description: List of ids to delete
      in: query
      required: true
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/IntegerIds'
...
    pathOrderId:
      name: orderId
      description: Limit results to specific order id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/IntegerId'
...
    employeeId:
      name: employeeId
      description: Limit results to specific employee id
      in: query
      schema:
        $ref: '#/components/schemas/IntegerId'
...
    optedServicesIds:
      name: optedServicesIds
      description: Limit result to matching optedServices ids
      in: query
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/IntegerIds'
...
    mimeTypes:
      name: mimeTypes
      description: Limit result set to specified mime types
      in: query
      style: form
      explode: false
      schema:
        type: array
        items:
          $ref: '#/components/schemas/MimeType'
...
    paginationInclude:
      name: include
      description: Get extra fields in result
      in: query
      style: form
      explode: false
      schema:
        type: array
        items:
          enum:
          - _pagination.total
...
    merchantInclude:
      name: include
      description: List of extra fields to include in response
      in: query
      style: form
      explode: false
      schema:
        type: array
        items:
          enum:
            - profile
            - salons
    merchantIncludeWithPagination:
      name: include
      description: List of extra fields to include in response
      in: query
      style: form
      explode: false
      schema:
        type: array
        items:
          enum:
            - profile
            - salons
            - _pagination.total
...
    page:
      name: page
      description: Current page no. for pagination
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    perPage:
      name: perPage
      description: No. of records in paginated list
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
...
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    '500':
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    PositiveIntegerIncludingZero:
      type: integer
      minimum: 0
    IntegerId:
      type: integer
      minimum: 1
    IntegerIds:
      type: array
      items:
        $ref: '#/components/schemas/IntegerId'
      minItems: 1
    ShortText:
      type: string
      minLength: 2
      maxLength: 64
    MediumText:
      type: string
      minLength: 2
      maxLength: 255
...
    Money:
      type: number
      format: float
      minimum: 0
    Mobile:
      type: string
      format: regex
      pattern: '^[6-9]\d{9}$'
...
    DateTime:
      type: string
      format: date-time
      description: ISO 8601 compliant date time
    FileType:
      enum:
        - image
    MimeType:
      enum:
        - image/jpeg
        - image/png
        - image/gif
        - image/webp
        - image/svg+xml
...
    Sex:
      enum:
        - male
        - female
        - other
...
    Role:
      enum:
        - customer
        - merchant
        - employee
        - admin
...
    File:
      type: object
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/IntegerId'
          readOnly: true
        name:
          allOf:
            - $ref: '#/components/schemas/MediumText'
          readOnly: true
        type:
          allOf:
            - $ref: '#/components/schemas/FileType'
          readOnly: true
        title:
          $ref: '#/components/schemas/MediumText'
        mimeType:
          allOf:
            - $ref: '#/components/schemas/MimeType'
          readOnly: true
        url:
          type: string
          format: uri
          readOnly: true
        thumbnailURL:
          type: string
          format: uri
          readOnly: true
        updatedAt:
          $ref: '#/components/schemas/DateTime'
      required:
        - title
        - type
        - url
...
    UserUpdateBody:
      type: object
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/IntegerId'
          readOnly: true
        firstName:
          $ref: '#/components/schemas/ShortText'
        lastName:
          $ref: '#/components/schemas/ShortText'
        gender:
          $ref: '#/components/schemas/Sex'
        mobile:
          $ref: '#/components/schemas/Mobile'
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          writeOnly: true
        email:
          type: string
          format: email
        role:
          allOf:
            - $ref: '#/components/schemas/Role'
          readOnly: true
        profileImage:
          allOf:
            - $ref: '#/components/schemas/File'
          readOnly: true
        profileImageId:
          $ref: '#/components/schemas/IntegerId'
    User:
      allOf:
        - $ref: '#/components/schemas/UserUpdateBody'
      required:
        - firstName
        - gender
        - mobile
...
    Merchant:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            userId:
              allOf:
                - $ref: '#/components/schemas/IntegerId'
              readOnly: true
            aadhar:
              type: string
              format: regex
              pattern: '^\d{12}$'
          required:
            - lastName
            - email
            - aadhar
...
    OrderItemUpdateBody:
      type: object
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/IntegerId'
          readOnly: true
        orderId:
          $ref: '#/components/schemas/IntegerId'
        optedServiceId:
          $ref: '#/components/schemas/IntegerId'
        employeeId:
          $ref: '#/components/schemas/IntegerId'
        price:
          $ref: '#/components/schemas/Money'
        note:
          type: string
    OrderItem:
      allOf:
        - $ref: '#/components/schemas/OrderItemUpdateBody'
      required:
        - optedServiceId
        - price
...
    Error:
      type: object
      properties:
        code:
          oneOf:
            - type: number
            - type: string
        message:
          type: string
        data:
          oneOf:
            - type: object
            - type: array
              items: {}
      required:
        - message
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        perPage:
          type: integer
          minimum: 1
        total:
          type: integer
      required:
        - page
        - perPage
    JSONResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/JSONResponse'
        - type: object
          properties:
            success:
              type: boolean
              default: true
    SingleSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
          required:
            - data
    ListSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items: {}
          required:
            - data
    PaginatedListSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/ListSuccessResponse'
        - type: object
          properties:
            _pagination:
              $ref: '#/components/schemas/Pagination'
          required:
            - _pagination
...
    MultiObjectOperationSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          default: true
        data:
          type: object
          properties:
            count:
              $ref: '#/components/schemas/PositiveIntegerIncludingZero'
          required:
            - count
      required:
        - success
        - data
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/JSONResponse'
        - type: object
          properties:
            error:
              $ref: '#/components/schemas/Error'
          required:
            - error
  examples:
...
    file:
      value:
        title: Some Image
...
    merchant:
      value:
        firstName: Vinod
        lastName: Sharma
        gender: male
        mobile: '9862898720'
        password: pa$$word
        email: vinod@sharma.com
        profileImageId: 1
        aadhar: '123456789012'
...
    orderItem:
      value:
        optedServiceId: 1
        employeeId: 1
        price: 20
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT